backend/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
curl_cffi==0.11.1
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
Django==4.2.21
django-cors-headers==4.7.0
django-crontab==0.7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
frozendict==2.4.6
fsspec==2024.9.0
h11==0.16.0
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lightgbm==4.5.0
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mnist==0.2.2
mpmath==1.3.0
multitasking==0.0.11
nest-asyncio==1.6.0
networkx==3.4.1
numpy==2.1.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parso==0.8.4
peewee==3.18.1
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==6.31.0
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
pywin32==307
pyzmq==26.2.0
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.33.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.7.0
torchvision==0.22.0
tornado==6.4.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2024.2
urllib3==2.4.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yfinance==0.2.61
